requirements.txt
Django~=3.1.4
streamlit~=0.87.0
Pillow~=6.2.1
pandas~=1.3.0
fastai~=1.0.61
pip~=21.2.4
wheel~=0.37.0
cryptography~=3.4.7
lxml~=4.5.0
setuptools~=57.4.0
torch~=1.4.0+cpu
fastcore~=1.3.25